uterm: vt: remove fake-VT SIGUSR1/2 (de)activation
This whole concept was broken from the beginning. With hotkey based
activation we have a much better debugging tool. This is still very
fragile, but better than nothing. And we are doing pretty well in error
recovery during hijacked VT switches so that's not as problematic as one
might think.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>