uterm: video: fbdev: send UTERM_NEW/GONE events for fbdev devices
We didn't send these events as the fake-display for fbdev video-objects
was created during video-creation and thus no listener could have been
registered so far.
However, this caused every video listener to iterate over all displays on
wake-up. This is really ugly and we want to avoid that. Therefore, we now
send the UTERM_NEW event in an idle-handler and the GONE event on
shutdown.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>