tsm: vte: implement alternate-screen modes
On DECSET 47, 1047-1049 we switch between normal screen mode and alternate
screen mode. xterm and friends support a titeInhibit resource that
disables this feature, so we add a flag to set this, too. However, there
is currently no way to set this on the command-line. If someone needs
that, we can easily add that.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>