seat: fix activating only enabled seats during session-switch
If we deactivate a session, we must not activate the next session we find
but we need to look for an enabled session instead.
This also changes the session_deactivate() helper to allow deactivating
disabled sessions. This shouldn't be relyed on, but it is a safe fallback
if some session turns out to be disabled but still active.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>