uterm: video: add display callbacks
Users might be interested in vertical-blank events so we now allow them to
register event-callbacks on displays. The only event that we currently
pass is a page-flip event that is always sent as reaction to a swap. If
there is no hardware support for such events, we simply throw it from
within the swap() function.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>