shl: dlist: add *_but_one() loops
The *_but_one() variants of dlist-loops start from within a loop instead
of at the head and iterate over all elements except the start-element. The
start element is skipped as we would need a temporary variable to include
it and in most times you want to skip the start-element, anyway.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>