copying: remove no longer relevent sections
All of the (xkb) keyboard handling code was written by David and me;
even the old code was not based on xlib but written against the
specification (not entirely correctly, too). All of this logic was moved
to libxkbcommon, which we marely link against.

We also no longer carray our own uterm_keysyms.h file, so that section is
not relevant as well.

Signed-off-by: Ran Benita <ran234@gmail.com>