seat: use dummy session
This makes the seat implementation register one dummy session as fallback.
This session isn't switched to except if no other session is available.
Furthermore, if the dummy is active and a new session is registered, it is
automatically activated.

External applications should never know of this dummy session. It's a
kmscon internal fallback so we can draw the screen even if there is no
other session available.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>