shl: dlist: fix *_but_one() loops to allow "start == head"
If we iterate over a non-empty list and pass the same pointer for start
and head, then we will never leave that loop.

It might be unclear why this is supported at all as we can simply use the
shl_dlist_for_each() for that. However, it makes code a lot easier to read
if the user can simply use shl_dlist_for_each_but_one() without checking
whether they passed the same for start and head.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>