seat: fix delayed VT switches
If we switch VT and the current session cannot be put asleep, and later it
notifies us that it went asleep, we should then switch VT and not
reschedule the sessions.

We do this by remembering whether a VT switch is pending. The new
uterm-retry infrastructure allows us to easily repeat VT switches then.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>