cdev: fix copying pollhandles correctly
The pollhandle is solely implemented in libfuse and does not correspond to
some kernel-intern object. Instead, the kernel notifies us when it has
poll-waiters via a flag. If that flag is set we should notify the kernel
whenever the poll-flags change.

Therefore, we can safely copy the pollhandle all the time and delete our
temporary handles.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>