conf: improve --vt warning by allowing --seats=,all
If a seat is called "all" and we want to use it as single seat, we
currently have to pass "--seats=all," or "--seats=,all" etc. so it is not
detected as --seats=all.
However, our warning when --vt is used with it does not respect this.
Improve this warning to detect this case and correctly ignore it then.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>