shl: hook: allow callbacks to destroy the hook
If we currently dispatch a callback and they unregister themself and then
destroy the hook, we currently do not free the hook as we would free
memory that is used in the callstack.
However, it is a valid use-case for hook-callbacks to destroy the hook.
Therefore, we set a "dead" flag and destroy the hook after all callbacks
are done.

This fixes some memory-leaks inside the shared-signal callbacks of eloop.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>